Pseudo- Auth + registration (#6)
- all links expect login beforehand
- login should happen for an already existing user. Saves logged user to sesion
- cakes displayed are owned or Public ones
- logout links clears the session

+++ grails-app/domain/cakecatalog/Cake.groovy
@@ -5,7 +5,7 @@ class Cake {
   long id
   String name
   String description
-  int ownerId
+  long ownerId
   boolean isPublic
 
   static constraints = {
